add: improve code compatibility and fix various 32-bit to 64-bit data type issues

+++ src/seenserv.c
@@ -289,12 +289,12 @@ BOOL seenserv_db_load(void) {
 									si->ip = seen32.ip;
 									si->mode = seen32.mode;
 									si->type = seen32.type;
-									si->nick = (STR) seen32.nick ;
-									si->username = (STR) (seen32.username);
-									si->host = (STR) (seen32.host);
-									si->realname = (STR) (seen32.realname);
-									si->tempnick = (STR) (seen32.tempnick);
-									si->quitmsg = (STR) (seen32.quitmsg);
+									si->nick = (STR)(uintptr_t) seen32.nick ;
+									si->username = (STR)(uintptr_t) (seen32.username);
+									si->host = (STR)(uintptr_t) (seen32.host);
+									si->realname = (STR)(uintptr_t) (seen32.realname);
+									si->tempnick = (STR)(uintptr_t) (seen32.tempnick);
+									si->quitmsg = (STR)(uintptr_t) (seen32.quitmsg);
 								}
 #else
 								result = stg_read_record(stg, (PBYTE)si, sizeof(SeenInfo_V10));
